requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
black==22.12.0
bootstrap4==0.1.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
Django==4.1.5
django-bootstrap4==22.3
django-environ==0.9.0
django-extensions==3.2.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.3
Pillow==9.4.0
pip==22.3.1
platformdirs==2.6.2
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.7.1
setuptools==65.5.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
wheel==0.37.1
